move session methods back into ClientMfaServer
