+++ crates/defguard_core/src/grpc/proxy/mod.rs
@@ -1,2 +1,1 @@
 pub mod client_mfa;
-pub mod session;
